# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiErrorBody": {
        "description": "错误体。不输出堆栈、SQL、密钥或完整供应商签名 URL（contracts.md §1）。",
        "properties": {
          "code": {
            "description": "机器可读错误码（SCREAMING_SNAKE_CASE，来自 `manual_core::ApiErrorCode`）。",
            "example": "NOT_FOUND",
            "type": "string"
          },
          "details": {
            "description": "结构化细节；无细节时为 null。"
          },
          "message": {
            "description": "面向用户的中文信息，不含内部细节。",
            "example": "接口不存在",
            "type": "string"
          },
          "requestId": {
            "description": "本次响应的诊断 ID（T04 起与请求日志关联）。",
            "example": "01993000-0000-7000-8000-000000000001",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "type": "object"
      },
      "ApiErrorResponse": {
        "description": "统一错误响应包。任何非 2xx 的应用错误都使用该结构。",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "PhotoArrangementRequest": {
        "additionalProperties": false,
        "properties": {
          "slots": {
            "additionalProperties": {
              "type": [
                "string",
                "null"
              ]
            },
            "description": "每个视图要放的照片资产；省略或 null = 该视图为空。资产须为本物品 `photo`，且同一资产不能放两个视图。",
            "propertyNames": {
              "type": "string"
            },
            "type": "object"
          }
        },
        "type": "object"
      },
...
      "PhotoDto": {
        "description": "photo 的线上表示。",
        "properties": {
          "assetId": {
            "description": "照片资产 ID（`purpose=photo`，JPEG/PNG）；内容经 `/assets/{id}/content` 读取。",
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "itemId": {
            "type": "string"
          },
          "revision": {
            "description": "乐观锁版本；GET 时对应 `ETag: \"r<revision>\"`。",
            "format": "int64",
            "type": "integer"
          },
          "updatedAt": {
            "type": "string"
          },
          "view": {
            "example": "front",
            "type": "string"
          }
        },
        "required": [
          "id",
          "itemId",
          "assetId",
          "view",
          "revision",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
      "PhotoListResponse": {
        "description": "照片列表响应。\n\n物品的照片集合被视图唯一性约束上界为 5 条（front/left/back/right/detail 各一张），\n因此不设分页：`nextCursor` 恒为 null（不是截断）。",
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/PhotoDto"
            },
            "type": "array"
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
...
      "ViewCandidateCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "assetId": {
            "description": "候选图资产（本物品、`purpose=photo`、JPEG/PNG）。",
            "type": "string"
          },
          "classify": {
            "description": "是否请说明书 AI 判断建议视图（默认 false，须显式确认图片发送与可能费用）。",
            "type": [
              "boolean",
              "null"
            ]
          },
          "documentId": {
            "type": [
              "string",
              "null"
            ]
          },
          "pageNumber": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "source": {
            "description": "`embedded` / `region` / `upload`。",
            "type": "string"
          }
        },
        "required": [
          "assetId",
          "source"
        ],
        "type": "object"
      },
      "ViewCandidateDto": {
        "properties": {
          "assetId": {
            "type": "string"
          },
          "confidence": {
            "format": "double",
            "type": [
              "number",
              "null"
            ]
          },
          "createdAt": {
            "type": "string"
          },
          "documentId": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "itemId": {
            "type": "string"
          },
          "note": {
            "description": "判断理由（简短，供用户核对）。",
            "type": [
              "string",
              "null"
            ]
          },
          "pageNumber": {
            "description": "来源页（1-based）；手动上传的候选为 null。",
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "source": {
            "description": "`embedded`（PDF 内嵌位图）/ `region`（渲染页的图形区域裁剪）/ `upload`（用户上传）。",
            "type": "string"
          },
          "suggestedView": {
            "description": "建议视图（front/left/back/right/detail）；null = 未判断或不像产品视图。",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "itemId",
          "assetId",
          "source",
          "createdAt"
        ],
        "type": "object"
      },
      "ViewCandidateListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ViewCandidateDto"
            },
            "type": "array"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
      "ViewCandidateResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ViewCandidateDto"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "in": "cookie",
        "name": "em_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "自托管交互说明书的本地 HTTP API。健康探针、认证/会话、设置状态、资产上传/内容服务（T06）、物品与说明书绑定/多视图照片（T07）、其余路由按 llmdoc/contracts.md 在同一文档上扩展。",
    "title": "万物说明书 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/items/{id}/photos/arrangement": {
      "put": {
        "description": "slots 给出 front/left/back/right/detail 各自的照片资产（省略/null = 空）。资产须为本物品照片且不重复。在一个写事务内替换本物品的全部照片行（同资产保留原照片 id），因此互换两个视图不会撞唯一索引。已开始任务使用冻结快照，不受影响；已有报价会因输入变化失效。",
        "operationId": "arrange_photos",
        "parameters": [
          {
            "description": "物品 ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhotoArrangementRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoListResponse"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "一次性确定全部视图照片（拖拽排列的保存）",
        "tags": [
          "photos"
        ]
      }
    },
...
    "/api/v1/items/{id}/view-candidates": {
      "get": {
        "operationId": "list_candidates",
        "parameters": [
          {
            "description": "物品 ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewCandidateListResponse"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "视图候选图（从说明书 PDF 拆出的待选图）",
        "tags": [
          "photos"
        ]
      },
      "post": {
        "description": "资产须属于本物品且为照片（JPEG/PNG）。同一资产重复登记返回已有候选（200，不重复判断）。classify 默认 false：仅显式 true 时调用已配置的说明书 AI（与提取同一服务端密钥），失败时候选记为未判断（仍 201），不阻塞用户手动排列。候选不进入报价/生成快照。",
        "operationId": "create_candidate",
        "parameters": [
          {
            "description": "物品 ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ViewCandidateCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewCandidateResponse"
                }
              }
            },
            "description": ""
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewCandidateResponse"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "登记一张视图候选图（可选由说明书 AI 判断建议视图）",
        "tags": [
          "photos"
        ]
      }
    },
    "/api/v1/items/{id}/view-candidates/{candidateId}/dismiss": {
      "post": {
        "operationId": "dismiss_candidate",
        "parameters": [
          {
            "description": "物品 ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "候选 ID",
            "in": "path",
            "name": "candidateId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "已删除"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "删除一张候选（软删除，可撤销；不影响已排列的照片）",
        "tags": [
          "photos"
        ]
      }
    },
    "/api/v1/items/{id}/view-candidates/{candidateId}/restore": {
      "post": {
        "operationId": "restore_candidate",
        "parameters": [
          {
            "description": "物品 ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "候选 ID",
            "in": "path",
            "name": "candidateId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "已恢复"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "撤销删除候选",
        "tags": [
          "photos"
        ]
      }
    },
...
  }
}